Revert dispatch types refactoring
