+++ src/cattrs/preconf/msgspec.py
@@ -17,11 +17,11 @@ from msgspec.json import Encoder, decode
 from .._compat import fields, get_args, get_origin, is_bare, is_mapping, is_sequence
 from ..cols import is_namedtuple
 from ..converters import BaseConverter, Converter
+from ..dispatch import UnstructureHook
 from ..fns import identity
 from ..gen import make_hetero_tuple_unstructure_fn
 from ..literals import is_literal_containing_enums
 from ..strategies import configure_union_passthrough
-from ..types import UnstructureHook
 from . import literals_with_enums_unstructure_factory, wrap
 
 T = TypeVar("T")
